pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project.urls]
repository = "https://github.com/TQRG/tenet"
homepage = "https://github.com/TQRG/tenet"

[project]
name="tenet"
version = "0.0.1"
description = "Fine-grained approach to detect and patch vulnerabilities"
readme = "README.md"
authors = [
    {name = "Eduard Pinconschi", email="eduard.pinconschi@tecnico.ulisboa.pt"},
    {name = "Sofia Reis", email="sofia.o.reis@tecnico.ulisboa.pt"}
]
keywords = ["machine learning", "security", "program analysis", "framework"]
license = {file = "LICENSE.txt"}
requires-python = ">=3.9"
classifiers = [
    "Topic :: Software Development",
    "Programming Language :: Python",
    "Environment :: Console",
    "Intended Audience :: Developers"
]
dependencies = [
    "cement==3.0.8",
    "jinja2",
    "pyyaml~=5.4.1",
    "colorlog",
    "pygithub~=1.55",
    "tqdm~=4.62.3",
    "gitpython",
    "beautifulsoup4",
    "pandas~=1.3.3",
    "jsbeautifier~=1.14.0",
    "requests~=2.26.0",
    "numpy~=1.21.2",
    "scikit-learn~=1.0",
    "imblearn~=0.0",
    "termcolor~=2.1.1",
    "docker~=5.0.3",
    "psutil~=5.9.3",
    "jsonlines~=2.0.0",
    "schema~=0.7.4",
    "matplotlib~=3.4.3",
    "seaborn~=0.11.2",
    "jq~=1.2.1",
    "pyparsing~=3.0.6",
    "code_tokenize~=0.2.0",
    "scipy~=1.9.2",
    "kaleido==0.2.1",
    "google-cloud-storage",
    "plotly~=5.11.0"
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "coverage",
    "twine>=1.11.0",
    "setuptools>=38.6.0",   
    "wheel>=0.31.0"
]


[tool.setuptools.packages.find]
exclude = ["ez_setup", "tests*"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
tenet = ["templates/*", "config/*"]

[project.scripts]
tenet = "tenet.main:main"